% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOptimalThreshold.R
\name{findOptimalFilter}
\alias{findOptimalFilter}
\title{Find the optimal filter threshold}
\usage{
findOptimalFilter(pvalues, filter, above = FALSE, threshold = 0.05)
}
\arguments{
\item{pvalues}{Numeric vector of p-values, one for each hypothesis.}

\item{filter}{Numeric vector of length equal to \code{pvalues}, 
containing the filter statistic for each hypothesis.}

\item{above}{Logical scalar indicating whether to retain hypotheses with \code{filter} above the threshold.
By default, lower values of \code{filter} are retained when choosing a filter statistic.}

\item{threshold}{Numeric scalar specifying the FDR threshold to use.}
}
\value{
A list containing \code{threshold}, the threshold to apply to \code{filter};
and \code{number}, the number of discoveries after applying the BH method at this filter threshold.
}
\description{
Find the threshold on an independent filter statistic that maximizes the number of discoveries
after applying the Benjamini-Hochberg method to control the false discovery rate.
}
\examples{
pvalues <- c(runif(9900), rbeta(100, 1, 50))
filter <- c(rnorm(9900), rnorm(100) + 2)
findOptimalFilter(pvalues, filter, above=TRUE)

}
\author{
Aaron Lun
}
